requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
isort==5.12.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
yarl==1.9.2
